Added live lyric display using conky (unfinished)
This uses another conky instance to display the lyrics of the
current song playing in Spotify using `sptlrx`.
However, one feature which also displays the song's title and artist
does not exactly function as I'd hoped. The overall function works, but
a problems arises with synchronization. Whenever a song changes,
`displaySongTitle` won't not update until `displaySongLyrics` updates
(i.e. changes to a new lyric line). So, if a song without any lyrics
available is played, `displaySongTitle` never updates as there aren't
any available lyrics for `displaySongLyrics` to display.
This will be fixed soon. (i really want it to work afterall)